create basic command and box DTO for calculations later
This command will read the files line by line and process it line by
line so less memory will be used on larger files.
This might seem as an overkill as a file_get_contents would suffice, but
I think for te sake of best practice this will be the best way.

The read data will be populated into a Box DTO object to isolate
claculations later on the box surface and perimeter.