feat: Calculate day 2 task 2 - ribbon size
The task says to calculate the total ribbon needed for the presents:
First we need the length of the ribbon which is the perimeter of the
smallest side.
For this we need the first and the second smallest edges of the box and
calculate the paremeter.

Then the ribbon length is just the volume of the box which is V=l*w*h